Update superblock with list of checkpoints
On opening the image it reads the last checkpoint in the list and then
rolls forward from that point, stopping at the first missing object, or
rolling forward from the beginning if there are no checkpoints. When you
forget to record the highest checkpoint, the low-numbered objects get
deleted by GC and all the data on the image gets lost.